+++ src/app/events/constants/sponsorTierConfigs.ts
